refactor: uninstall webpack
