+++ index.html
@@ -12,6 +12,6 @@
         <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
     </head>
     <body>
-        <script src="script.js" defer></script>
+        <script type="module" src="script.js" defer></script>
     </body>
 </html>
