+++ package.json
@@ -3,12 +3,9 @@
   "version": "1.0.0",
   "description": "",
   "private": true,
-  "main": "index.js",
+  "main": "script.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "webpack --mode development",
-    "build": "webpack --mode production",
-    "start": "webpack --mode production --watch",
     "eslint": "eslint"
   },
   "repository": {
@@ -23,15 +20,8 @@
   },
   "homepage": "https://github.com/sylarBrest/virtual-keyboard#readme",
   "devDependencies": {
-    "css-loader": "^6.7.1",
     "eslint": "^8.14.0",
     "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-webpack-plugin": "^3.1.1",
-    "mini-css-extract-plugin": "^2.6.0",
-    "node-sass": "^7.0.1",
-    "sass-loader": "^12.6.0",
-    "webpack": "^5.72.0",
-    "webpack-cli": "^4.9.2"
+    "eslint-plugin-import": "^2.26.0"
   }
 }
