--- public/style.css
@@ -1,2 +1,0 @@
-*{margin:0;padding:0;box-sizing:border-box}body{height:100vh;width:100vw;background-color:#cecece;font-family:monospace;font-size:20px;color:black}.main-container{max-width:778px;width:100%;height:auto;margin:0 auto;padding:20px 40px;display:flex;flex-flow:column;align-items:center;gap:20px;justify-content:center}.main-container .textarea{width:100%;margin:0 auto;padding:5px;font:inherit;resize:none;outline:0}.main-container .keyboard{display:flex;width:698px;gap:7px;flex-wrap:wrap;justify-content:center}.main-container .keyboard .key{height:40px;width:40px;display:flex;justify-content:center;align-items:center;font:inherit;font-family:sans-serif;cursor:pointer;transition:0.3s all;border:0;border-radius:5px}.main-container .keyboard .key.wide{width:87px}.main-container .keyboard .key.ultrawide{width:322px}.main-container .keyboard .key.text{font-size:16px;font-weight:700}.main-container .keyboard .key:hover{background:#c7c7c7}.main-container .keyboard .key:active{background:#838383;box-shadow:inset 0 0 20px #292929}.main-container .keyboard .key.pressed{background:#838383;box-shadow:inset 0 0 20px #292929}
-
