+++ script.js
@@ -1,4 +1,4 @@
-export default class Keyboard {
+class Keyboard {
   constructor() {
     this.elements = {
       main: null,
@@ -391,3 +391,26 @@ export default class Keyboard {
     this.reLoadKeys(keys);
   }
 }
+
+/* TODO or not TODO List
+1. Shift, CapsLock and LangSwitching only after keyup event
+*/
+
+console.log('Virtual Keyboard task started');
+
+const kb = new Keyboard();
+
+const setLocalStorage = () => localStorage.setItem('lang', kb.getLang());
+
+const getLocalStorage = () => {
+  if (localStorage.getItem('lang')) {
+    kb.init();
+    kb.setLang(localStorage.getItem('lang'));
+  } else {
+    setLocalStorage();
+    getLocalStorage();
+  }
+};
+
+window.addEventListener('load', getLocalStorage);
+window.addEventListener('beforeunload', setLocalStorage);
